testing/web-platform/mozilla/tests/dom/events/retrieve_coords_after_moving_target_outside_subdoc_and_hide_the_subdoc.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<title>Retrieve coordinates after event target is moved to outside the subdocument and hide the subdocument</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script>
"use strict";

/**
 * This is a copy of
 * uievents/mouse/crashtests/retrieve_coords_after_moving_target_outside_subdoc_and_hide_the_subdoc.html
 * It's skipped only on Firefox because of using test driver in a crash test.
 * So, once it's fixed, this won't be required.
 */

addEventListener("load", () => {
  promise_test(async () => {
    const iframe = document.querySelector("iframe");
    const div = iframe.contentDocument.getElementById("target");
    let clickEvent;
    div.addEventListener("click", event => clickEvent = event, {once: true});
    await test_driver.click(div);
    document.body.appendChild(div);
    iframe.hidden = true;
    await new Promise(resolve => requestAnimationFrame(() => requestAnimationFrame(resolve)));
    div.innerText = clickEvent.screenX;
    div.innerText = clickEvent.screenY;
    div.innerText = clickEvent.clientX;
    div.innerText = clickEvent.clientY;
    div.innerText = clickEvent.offsetX;
    div.innerText = clickEvent.offsetY;
    div.innerText = clickEvent.pageX;
    div.innerText = clickEvent.pageY;
    div.innerText = clickEvent.layerX;
    div.innerText = clickEvent.layerY;
    assert_true(true, "Shouldn't crash");
  });
}, {once: true});
</script>
</head>
<body>
  <iframe srcdoc="<body><div id='target' style='height:32px'></div></body>"></iframe>
</body>
</html>
